--- src/main/java/com/backend/immilog/image/domain/enums/ImageStatus.kt
@@ -1,6 +1,0 @@
-package com.backend.immilog.image.domain.enums
-
-enum class ImageStatus {
-    NORMAL,
-    DELETED
-}
