--- src/main/java/com/backend/immilog/image/domain/enums/ImageType.kt
@@ -1,7 +1,0 @@
-package com.backend.immilog.image.domain.enums
-
-enum class ImageType {
-    PROFILE,
-    POST,
-    JOB_POST
-}
