+++ src/main/java/com/backend/immilog/image/infrastructure/persistence/repository/ImageRepositoryImpl.kt
@@ -1,11 +1,11 @@
 package com.backend.immilog.image.infrastructure.persistence.repository
 
-import com.backend.immilog.image.domain.model.Image
-import com.backend.immilog.image.domain.repository.ImageRepository
+import com.backend.immilog.image.domain.Image
+import com.backend.immilog.image.domain.ImageRepository
 import com.backend.immilog.image.exception.ImageErrorCode
 import com.backend.immilog.image.exception.ImageException
-import com.backend.immilog.image.infrastructure.persistence.jpa.entity.ImageEntity
-import com.backend.immilog.image.infrastructure.persistence.jpa.repository.ImageJpaRepository
+import com.backend.immilog.image.infrastructure.persistence.jpa.ImageJpaEntity
+import com.backend.immilog.image.infrastructure.persistence.jpa.ImageJpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
@@ -13,7 +13,7 @@ class ImageRepositoryImpl(
     private val imageJpaRepository: ImageJpaRepository
 ) : ImageRepository {
     override fun save(image: Image): Image {
-        return imageJpaRepository.save(ImageEntity.from(image)).toDomain()
+        return imageJpaRepository.save(ImageJpaEntity.from(image)).toDomain()
     }
 
     override fun findByPath(imagePath: String): Image {
