+++ src/main/java/com/backend/immilog/image/presentation/ImageResponse.kt
@@ -1,4 +1,4 @@
-package com.backend.immilog.image.presentation.payload
+package com.backend.immilog.image.presentation
 
 import io.swagger.v3.oas.annotations.media.Schema
 import org.springframework.http.HttpStatus
